refactor: extract handleNormalKey into normal.ts
