+++ src/vim/index.ts
@@ -1,341 +1,17 @@
-import { consumeCount, enterInsert, enterNormal, exitVisual, resetPending } from "./state";
-import { DELETE_MOTION, MOTIONS, SELECT_MOTIONS } from "./tables";
-import { currentLineRange, endOfWord } from "./text";
+import { consumeCount, enterInsert, enterNormal, exitVisual } from "./state";
+import { SELECT_MOTIONS } from "./tables";
+import { endOfWord } from "./text";
 import type { Action, HandlerResult, KeyEvent, PromptAccess, VimState } from "./types";
 import { PASS, pushN } from "./util";
 
 export { handleInsertKey } from "./insert";
+export { handleNormalKey } from "./normal";
 export { createVimState, finishOneShotIfComplete, toggleVimMode } from "./state";
 export { endOfWord } from "./text";
 export type { Action, KeyEvent, PromptAccess, VimState } from "./types";
 export { translateKey } from "./util";
 
 const _CONSUME: HandlerResult = { consume: true, actions: [] };
-
-export function handleNormalKey(state: VimState, key: string, ev: KeyEvent, prompt: PromptAccess): HandlerResult {
-  if (ev.meta || ev.super) return PASS;
-  if (ev.ctrl) {
-    if (ev.name === "r") {
-      resetPending(state);
-      return { consume: true, actions: [{ type: "cmd", cmd: "input.redo" }] };
-    }
-    return PASS;
-  }
-
-  if (ev.name === "escape") {
-    if (state.oneShotNormal) {
-      state.oneShotNormal = false;
-      state.mode = "insert";
-      resetPending(state);
-      return { consume: true, actions: [{ type: "mode", mode: "insert" }] };
-    }
-    resetPending(state);
-    return PASS;
-  }
-
-  // Pending character argument (r{char})
-  if (state.pendingChar === "r") {
-    const n = consumeCount(state);
-    const actions: Action[] = [];
-    pushN(actions, "input.delete", n);
-    actions.push({ type: "insertText", text: key.repeat(n) });
-    state.pendingChar = null;
-    return finishUndoableChange(actions);
-  }
-
-  // Pending g prefix (gg, ge, etc.)
-  if (state.pendingChar === "g") {
-    state.pendingChar = null;
-    const actions: Action[] = [];
-    if (key === "g") {
-      consumeCount(state);
-      actions.push({ type: "cursorTo", offset: 0 });
-    } else {
-      resetPending(state);
-    }
-    return { consume: true, actions };
-  }
-
-  if (ev.name === "tab") return PASS;
-
-  // Everything below is consumed
-  const actions: Action[] = [];
-
-  if (/[1-9]/.test(key) || (key === "0" && state.count > 0)) {
-    state.count = state.count * 10 + parseInt(key, 10);
-    return { consume: true, actions };
-  }
-
-  if (ev.name === "return") {
-    actions.push({ type: "cmd", cmd: "input.submit" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === ":") {
-    actions.push({ type: "cmd", cmd: "command.palette.show" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "/") {
-    actions.push({ type: "cmd", cmd: "session.timeline" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "[") {
-    actions.push({ type: "cmd", cmd: "session.half.page.up" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "]") {
-    actions.push({ type: "cmd", cmd: "session.half.page.down" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "{") {
-    actions.push({ type: "cmd", cmd: "session.message.previous" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "}") {
-    actions.push({ type: "cmd", cmd: "session.message.next" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  if (key === "p") {
-    if (state.yankRegister) actions.push({ type: "yank", text: state.yankRegister });
-    actions.push({ type: "cmd", cmd: "prompt.paste" });
-    resetPending(state);
-    return finishUndoableChange(actions);
-  }
-
-  if (key === "X") {
-    pushN(actions, "input.backspace", consumeCount(state));
-    return finishUndoableChange(actions);
-  }
-
-  if (key === "J") {
-    const n = consumeCount(state);
-    for (let i = 0; i < n; i++) {
-      actions.push({ type: "cmd", cmd: "input.line.end" });
-      actions.push({ type: "cmd", cmd: "input.delete" });
-    }
-    return finishUndoableChange(actions);
-  }
-
-  // Operators: d, c, y
-  if (key === "d" || key === "c" || key === "y") {
-    if (state.pendingOp === key) {
-      const n = consumeCount(state);
-      if (key === "y") {
-        const cursorLine = prompt.getCursorLine();
-        const lines: string[] = [];
-        for (let i = 0; i < n; i++) lines.push(prompt.getLine(cursorLine + i));
-        const text = `${lines.join("\n")}\n`;
-        state.yankRegister = text;
-        actions.push({ type: "yank", text });
-        actions.push({ type: "toast", message: `${n} line${n > 1 ? "s" : ""} yanked`, duration: 1000 });
-        resetPending(state);
-      } else {
-        pushN(actions, "input.delete.line", n);
-        if (key === "c") enterInsert(state, actions);
-        else resetPending(state);
-        return finishUndoableChange(actions);
-      }
-      return { consume: true, actions };
-    }
-    state.pendingOp = key;
-    return { consume: true, actions };
-  }
-
-  if (key === "D") {
-    actions.push({ type: "cmd", cmd: "input.delete.to.line.end" });
-    resetPending(state);
-    return finishUndoableChange(actions);
-  }
-
-  if (key === "C") {
-    actions.push({ type: "cmd", cmd: "input.delete.to.line.end" });
-    enterInsert(state, actions);
-    return finishUndoableChange(actions);
-  }
-
-  // Pending operator + e (end-of-word needs special handling)
-  if (state.pendingOp && key === "e") {
-    const n = consumeCount(state);
-    const offset = prompt.getCursorOffset();
-    const target = endOfWord(prompt.getPlainText(), offset, n);
-    if (state.pendingOp === "y") {
-      const text = prompt.getPlainText().slice(offset, target + 1);
-      state.yankRegister = text;
-      actions.push({ type: "yank", text });
-      resetPending(state);
-      return { consume: true, actions };
-    }
-    actions.push({ type: "deleteRange", start: offset, end: target });
-    if (state.pendingOp === "c") enterInsert(state, actions);
-    else resetPending(state);
-    return finishUndoableChange(actions);
-  }
-
-  // Pending operator + motion
-  if (state.pendingOp && key in MOTIONS) {
-    const n = consumeCount(state);
-
-    if (state.pendingOp === "y") {
-      const selectCmd = SELECT_MOTIONS[key];
-      if (selectCmd) {
-        pushN(actions, selectCmd, n);
-        actions.push({ type: "yankSelection" });
-      }
-      resetPending(state);
-      return { consume: true, actions };
-    }
-
-    if (key === "j") {
-      pushN(actions, "input.delete.line", n + 1);
-      if (state.pendingOp === "c") enterInsert(state, actions);
-      else resetPending(state);
-      return finishUndoableChange(actions);
-    }
-    if (key === "k") {
-      pushN(actions, "input.move.up", n);
-      pushN(actions, "input.delete.line", n + 1);
-      if (state.pendingOp === "c") enterInsert(state, actions);
-      else resetPending(state);
-      return finishUndoableChange(actions);
-    }
-    if (key === "G") {
-      consumeCount(state);
-      const offset = prompt.getCursorOffset();
-      const text = prompt.getPlainText();
-      actions.push({ type: "deleteRange", start: offset, end: Math.max(0, text.length - 1) });
-      if (state.pendingOp === "c") enterInsert(state, actions);
-      else resetPending(state);
-      return finishUndoableChange(actions);
-    }
-
-    const deleteCmd = DELETE_MOTION[key];
-    if (deleteCmd) {
-      pushN(actions, deleteCmd, n);
-      if (state.pendingOp === "c") enterInsert(state, actions);
-      else resetPending(state);
-      return finishUndoableChange(actions);
-    }
-
-    // unreachable: every MOTIONS key reaching here has a DELETE_MOTION entry (j/k/G handled above)
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  // Standalone e (end-of-word)
-  if (key === "e") {
-    const n = consumeCount(state);
-    const target = endOfWord(prompt.getPlainText(), prompt.getCursorOffset(), n);
-    actions.push({ type: "cursorTo", offset: target });
-    return { consume: true, actions };
-  }
-
-  // Standalone motions
-  if (key in MOTIONS) {
-    const n = consumeCount(state);
-    if ((key === "j" || key === "k") && isInputEmpty(prompt)) {
-      const cmd = key === "k" ? "prompt.history.previous" : "prompt.history.next";
-      pushN(actions, cmd, n);
-      return { consume: true, actions };
-    }
-    pushN(actions, MOTIONS[key], n);
-    return { consume: true, actions };
-  }
-
-  // g prefix — wait for second keypress
-  if (key === "g") {
-    state.pendingChar = "g";
-    return { consume: true, actions };
-  }
-
-  if (key === "x") {
-    pushN(actions, "input.delete", consumeCount(state));
-    return finishUndoableChange(actions);
-  }
-
-  if (key === "r") {
-    state.pendingChar = "r";
-    return { consume: true, actions };
-  }
-
-  if (key === "u") {
-    actions.push({ type: "undo" });
-    resetPending(state);
-    return { consume: true, actions };
-  }
-
-  // Visual mode entry
-  if (key === "V") {
-    const range = currentLineRange(prompt.getPlainText(), prompt.getCursorOffset());
-    state.mode = "visual";
-    state.visualAnchor = prompt.getCursorOffset();
-    state.oneShotNormal = false;
-    resetPending(state);
-    return {
-      consume: true,
-      actions: [
-        { type: "selectRange", start: range.start, end: range.end },
-        { type: "mode", mode: "visual" },
-      ],
-    };
-  }
-
-  if (key === "v") {
-    state.mode = "visual";
-    state.visualAnchor = prompt.getCursorOffset();
-    state.oneShotNormal = false;
-    resetPending(state);
-    return { consume: true, actions: [{ type: "mode", mode: "visual" }] };
-  }
-
-  // Insert entries
-  if (key === "i") {
-    enterInsert(state, actions);
-    return { consume: true, actions };
-  }
-
-  if (key === "a") {
-    actions.push({ type: "cmd", cmd: "input.move.right" });
-    enterInsert(state, actions);
-    return { consume: true, actions };
-  }
-
-  if (key === "A") {
-    actions.push({ type: "cmd", cmd: "input.line.end" });
-    enterInsert(state, actions);
-    return { consume: true, actions };
-  }
-
-  if (key === "o") {
-    actions.push({ type: "cmd", cmd: "input.line.end" });
-    actions.push({ type: "cmd", cmd: "input.newline" });
-    enterInsert(state, actions);
-    return { consume: true, actions };
-  }
-
-  if (key === "O") {
-    actions.push({ type: "cmd", cmd: "input.line.home" });
-    actions.push({ type: "cmd", cmd: "input.newline" });
-    actions.push({ type: "cmd", cmd: "input.move.up" });
-    enterInsert(state, actions);
-    return { consume: true, actions };
-  }
-
-  // Unbound key — already consumed
-  return { consume: true, actions };
-}
 
 export function handleVisualKey(state: VimState, key: string, ev: KeyEvent, prompt: PromptAccess): HandlerResult {
   if (ev.meta || ev.super) return PASS;
@@ -408,14 +84,4 @@ export function handleVisualKey(state: VimState, key: string, ev: KeyEvent, prom
 
   // Unbound key — consume to prevent typing
   return { consume: true, actions };
-}
-
-// ── Helpers ──────────────────────────────────────────────────
-
-function finishUndoableChange(actions: Action[]): HandlerResult {
-  return { consume: true, actions: [{ type: "saveUndoSnapshot" }, ...actions] };
-}
-
-function isInputEmpty(prompt: PromptAccess): boolean {
-  return prompt.getLineCount() === 1 && prompt.getLine(0) === "";
 }
